+++ ECommerce/ViewModels/ProductViewModel.cs
@@ -4,9 +4,9 @@ public abstract record BaseProductViewModel
 {
     public string Title { get; set; }
     public string Description { get; set; }
-    public ushort Price { get; set; }
+    public double Price { get; set; }
 }
 
-public record ProductViewModel(string CategoryName, int Id): BaseProductViewModel;
+public record ProductViewModel(string CategoryName, int Id) : BaseProductViewModel;
 
-public record CreateProductViewModel(int CategoryId): BaseProductViewModel;
+public record CreateProductViewModel(int CategoryId) : BaseProductViewModel;
